Changing Doxyfile so that doxygen will scan files matching the pattern *.dox.
Also adding files src/README.dox indicating how to add documentation
to the "main page", and src/mpi/README.dox as an example of adding
a page of supplementary information.

Don't know if there is a "standard" way of doing this, but it seems
to work, and saves us having to stuff too much detail into real
source files.


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@964 63e3feb5-37d5-0310-a306-e8a459e722fe