Add the name server to the system. Code checked by Jeff - thanks!
The name server provides a 64-bit name - no structure to it. If I put a structure in (as Graham had proposed), then I had to track the allocations in each piece of the structure. Quite doable - but this was much simpler and adequate for now. Hard to believe we would hit 64-bits worth of names...but we can change it later, if needed.



git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@1302 63e3feb5-37d5-0310-a306-e8a459e722fe