Major change to the session directory system that will break all code that uses it - so BEWARE! I modified proc_info.c and universe_init.c to work with the new system, but those were the only places I could immediately find. Check out the src/util/session_dir.h documentation to see the changes.
The tree seems to still compile as far as I can tell. I can't do a full commit against the tree because someone made changes to the registry functions, thereby creating a conflict that I haven't had the chance to resolve. I believe I have committed all the required files, but we'll have to see...



git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@1959 63e3feb5-37d5-0310-a306-e8a459e722fe