Major update to the general purpose registry. Cleaned up the mess from name changes, begin building the functionality. Long way to go....
I have to commit this to cleanup a break in my tree. I'm hoping it won't break the compile of the tree, but will fix it as quickly as possible.

Jeff - you are welcome to set an "ignore" on the gpr if you like - I'll let you know when I've got the "kinks" out.



git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@2145 63e3feb5-37d5-0310-a306-e8a459e722fe