Workaround the mca_oob_ping problem by doing rapid multiple checks - works just fine.
We now have the ability to generate and join a persistent universe. You can create one in two ways:

(a) issue the "openmpi" command. This will fork/exec a seed daemon on your local host. You can specify a universe name or else it will just use the default.

(b) issue the "ompid -seed" command. Starts the seed up directly. Takes all the same options as openmpi.

I will be adjusting mpirun2 and mpi_init to allow connection to existing persistent universes, but they don't do it right now. The ompiconsole program simply issues an exit command to the persistent universe, so you can use it to shut the universe down if you like (or a kill -9  - works too).




git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@2629 63e3feb5-37d5-0310-a306-e8a459e722fe