Discovered a very, very interesting bug, courtesy of Jeff point out some unexpected behavior under Linux. It turns out that the Linux "dirname" function has a bug in it - namely, it modifies the argument passed to it, in violation of its documentation. This was apparently fixed in the Mac library - it definitely does NOT modify its argument on the Mac.
After spending hours tracking this down, I finally did manage to isolate it to the dirname function. I have installed a "fix" for the problem (using an intermediate temporary copy of the string being passed to the function) that solves the problem without messing us up on the Mac.

Hopefully, we won't find too many more such "cute" errors!


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@2643 63e3feb5-37d5-0310-a306-e8a459e722fe