Cleanup a bunch of warnings. Most were innocent enough (things not initialized that probably wouldn't be a problem), but a couple could be problematic.
One warning on my list was left unfixed: the "mca_ptl_tcp_peer_dump" function is defined but not used. It is commented in the file as a diagnostic, so I left it alone since someone may choose to use it for debugging purposes. For now, it is uncalled anywhere in the code.


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@3703 63e3feb5-37d5-0310-a306-e8a459e722fe