First cut at non-blocking IO progress.  No asychronous progress support is included because ROMIO is inherently thread-unsafe.
One possible way to have true asynchronous progress would be to use a
progress thread that wakes up and polls at some frequency when there
are non-blocking IO requests pending.  This is pretty icky, though --
it should definitely have an MCA parameter to enable/disable this
functionality, as well as another to control the polling frequency.

This also strengthens the argument that we need a v2 of the io
framework -- one that is not designed to exclusively support ROMIO --
one that does something unimaginably "better" for the parallel MPI-2
IO interface.  :-)


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@3786 63e3feb5-37d5-0310-a306-e8a459e722fe