clean up the OMPI_BUILDING #define.  Rather than being defined to 1 if we are part of the source tree and not defined otherwise, we are going with an always defined if ompi_config.h is included policy.  If ompi_config.h is included before mpi.h or before OMPI_BUILDING is set, it will set OMPI_BUILDING to 1 and enable all the internal code that is in ompi_config_bottom.h.  Otherwise, it will only include the system configuration data (enough for defining the C and C++ interfaces to MPI, but not perturbing the user environment).
This should fix the problems with bool and the like that the Eclipse
folks were seeing.  It also cleans up some build system hacks that
we had along the way.

Also, don't use int64_t as the default size of MPI_Offset, because it
requires us including stdint.h in mpi.h, which is something we really
shouldn't be doing.

And finally, fix a ROMIO Makefile that didn't set -DOMPI_BUILDING=1,
as ROMIO includes mpi.h, but not ompi_config.h


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@5430 63e3feb5-37d5-0310-a306-e8a459e722fe