Temporary fix -- will be obviated by the size_t (et al.) fixes coming soon.
Make ORTE_EXIT_CODE be the same as INT32, not INT8.  This allows the
full propogation of the value returned by waitpid() rather than just
the lowest 8 bites.  Also change the naming of it in orterun to be
exit_status, not exit_code (per POSIX standard naming convention).

orterun now returns the first nonzero exit status that it receives.


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@5530 63e3feb5-37d5-0310-a306-e8a459e722fe