Fix a bug in the new gpr match check function that caused some containers to be mismatched.
Add the logic to properly assign new cellid's to hosts read in by the hostfile component. However, don't turn it on yet.

It seems that the code base has (unfortunately) assumed that cellid is always zero. When I turn on the cellid capability, the system "hangs" whenever the cellid is non-zero. I'll have to chase that problem down. For now, I've turned "off" the cellid assignment in the hostfile component.


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@5865 63e3feb5-37d5-0310-a306-e8a459e722fe