Fix a problem found by Tim Prins (patch also supplied by Tim P).  From his e-mail:
I ran into a small bug in rmaps_rr.c: map_app_by_slot which was
triggered by using multiple app contexts. Basically, if not all the
slots we allocated on a node were used by an app, we would
automatically move onto the next node. This caused a problem with
multiple app contexts when the first app takes a partial allocation of
a node, the second app would not be able to access these slots because
we had already move past the node, and the byslot routine does not
wrap back around the list.


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@6766 63e3feb5-37d5-0310-a306-e8a459e722fe