Last commit to make the threaded case compiling without warnings. Next step try to make it working ...
Correct the spring of the vpid problem (similar to the one in the SM PTL).

Add one more argument to the MCA_BTL_SM_FIFO_WRITE macro who will get passed down to the 
MCA_BTL_SM_SIGNAL_PEER macro to allow it to have the fifo_fd file descriptor.



git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@7305 63e3feb5-37d5-0310-a306-e8a459e722fe