That was a bad*bad typo. To compute the flags I have to use & not |.
In the case where the user want to roll-back a convertor we should first rewind the
convertor to the beginning and then move it to the requested position. Otherwise the
logic do roll-back a convertor in an optimized way is really difficult. BTW, in terms of
performances one would prefer to have 2 convertors and play with position inside, than
just having one and roll it back.


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@7378 63e3feb5-37d5-0310-a306-e8a459e722fe