Some Fortran compilers actually will return that a type exists even if it doesn't support it -- the compiler will automatically convert the unsupported type to a type that it *does* support.  For example, if you try to use INTEGER*16 and the compiler doesn't support it, it may well automatically convert it to INTEGER*8 for you (!).  So we have to check the actual size of the type once we determine that the compiler doesn't error if we try to use it (i.e,. the compiler *might* support that type).  If the size doesn't match the expected size, then the compiler doesn't really support it.
The F77 configure code actually handled this properly.  The F90 code
did not quite do it right.  This patch brings the F90 code up to the
same structure as the F77 code, albiet not m4-ized properly.  I also
added a comment to config/f77_check.m4 that explains *why* we do this
extra size check (because no explanation was given).

The impetus for this was that xlf* on OS X 10.3 was not recognizing
that INTEGER*16 was not supported, and mpi-f90-interfaces.h was being
assembled incorrectly.  This patch fixes this problem.

There is still one more problem, but waiting for some help from Craig
R on that (function pointers in F90 declarations).


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@8107 63e3feb5-37d5-0310-a306-e8a459e722fe