fixing the fortran interface for MPI_Get_address. Since the address argument is defined to be
integer (KIND=MPI_ADDRESS_KIND)

we have to map that to an MPI_Aint in the f->c interface, else the
typecast back from the C to the Fortran interface will be wrong as well.
The same holds (at least) for all other new MPI-2 datatype functions
which take a byte-displacement or an address as an argument.



git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@8243 63e3feb5-37d5-0310-a306-e8a459e722fe