Removed all of the LAM stuff.
This needs to be gone over a few more times before it is allowed to see
daylight, but has come a long way.  Some sections may be off more than a little,
but the general idea is there.

Need to audit to make sure we don't call the ORTE VHNP's daemons :)



git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@9078 63e3feb5-37d5-0310-a306-e8a459e722fe