I did it under pressure !!!
The free lst using atomic operations. I didn't want to completely
change the behavior, so we still use a mutex for the extreme cases (like
no more available items and we cannot allocate more). I test it for a
while on non multi-threading environment, but not enough on a multi-threaded
build.


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@9623 63e3feb5-37d5-0310-a306-e8a459e722fe