This commit contains the necessary changes to get "mpirun a.out" working correctly with MPI_Comm_spawn.
The problem wiht MPI_Comm_spawn was that the 'parent' process was 
rmgr.create'ing and then rmgr.launch'ing the children via the rmgr proxy
component. The HNP saw these commands and processed them normally, but
since we never went through the HNP's rmgr (urm component) spawn() 
logic the triggers and key/value pairs were never created. So the
children were launched correctly, but since the HNP did not
have any triggers setup, never triggered the xcast for the
children to finish orte_init().

This fix puts the trigger and key/value pair initialization in 
rmgr_urm_spawn() for the 'mpirun a.out' case, *and* in the 
rmgr_base_unpack routine that deals with the creation of the
job for the child as requested by the proxy component. This
will allow the triggers to be registered for the proxy's request
which only happens during MPI_Comm_spawn*

Small change for a lot of debugging. Notice that his reverts r11037
to its previous version, and adds a newline to handle the spawn
cases.




git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@11046 63e3feb5-37d5-0310-a306-e8a459e722fe