MPI_COMM_GET_NAME had the compiler-added extra string length parameter (but didn't use it), but MPI_TYPE_GET_NAME and MPI_WIN_GET_NAME did not.
This commit changes all three functions to pass the compile-added
string length parameter to clear out the remainder of the string with
spaces (i.e., the rest of the string that was not set with the name).
This is what was done in LAM/MPI, and apparently what was done in
Sun's MPI, because the test that Rolf attached now passes.

Fixes #274.


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@11301 63e3feb5-37d5-0310-a306-e8a459e722fe